requirements.txt
aiohttp==3.12.6
numpy==2.2.6
python-socketio==5.13.0
python-engineio==4.12.1
